node_map.
